auido_prepare/src/requirements.txt
numpy
scipy
grpcio
grpcio-tools